Complete full CRUD functionality to users and tasks on the server

+++ Controllers/tasksController.cs
@@ -2,6 +2,9 @@ using Microsoft.AspNetCore.Mvc;
 using taskList.Services;
 using taskList.Models;
 using taskList.Interfaces;
+using Microsoft.AspNetCore.Authorization;
+using taskList.Controllers;
+
 
 
 namespace taskList.Controllers;
@@ -14,37 +17,48 @@ public class TaskController : ControllerBase
     IMyTaskService MyTaskService;
 
     public TaskController(IMyTaskService MyTaskService)
-        {
-            this.MyTaskService = MyTaskService;
-        }
+    {
+        this.MyTaskService = MyTaskService;
+    }
 
     [HttpGet]
+    [Authorize(Policy = "user")]
     public ActionResult<List<MyTask>> Get()
     {
-        return MyTaskService.GetAll();
+        return MyTaskService.GetAll(int.Parse(User.FindFirst("id")?.Value));
     }
 
     [HttpGet("{id}")]
+    [Authorize(Policy = "user")]
     public ActionResult<MyTask> Get(int id)
     {
         var task = MyTaskService.GetById(id);
         if (task == null)
             return NotFound();
+        if (!chekAuthorization(id))
+            return Unauthorized();
+
         return task;
     }
 
-            [HttpPost] 
-        public IActionResult Create(MyTask myTask)
-        {
-            MyTaskService.Add(myTask);
-            return CreatedAtAction(nameof(Create), new {id=myTask.Id}, myTask);
+    [HttpPost]
+    [Authorize(Policy = "user")]
+    public IActionResult Create(MyTask myTask)
+    {
+        myTask.Owner = int.Parse(User.FindFirst("Id").Value);
+        MyTaskService.Add(myTask);
+        return CreatedAtAction(nameof(Create), new { id = myTask.Id }, myTask);
 
-        }
+    }
 
 
     [HttpPut("{id}")]
+    [Authorize(Policy = "user")]
     public ActionResult Put(int id, MyTask newTask)
     {
+        if(!chekAuthorization(id))
+            return Unauthorized();
+        newTask.Owner = int.Parse(User.FindFirst("Id").Value);
         var result = MyTaskService.Update(id, newTask);
         if (!result)
         {
@@ -54,15 +68,23 @@ public class TaskController : ControllerBase
     }
 
     [HttpDelete("{id}")]
+    [Authorize(Policy = "user")]
     public ActionResult Delete(int id)
     {
         var task = MyTaskService.GetById(id);
-        if (task==null)
+        if (task == null)
             return NotFound();
-
+        if (!chekAuthorization(id))
+            return Unauthorized();
         MyTaskService.Delete(id);
 
         return NoContent();
+
+    }
+
+    private bool chekAuthorization(int taskId)
+    {
+        return MyTaskService.GetById(taskId).Owner == int.Parse(User.FindFirst("id").Value);
 
     }
 }
